Predictive model - Random forest Regression
The Global Findex database is the world’s most comprehensive data set on how adults save, borrow, make payments, and manage risk. Launched with funding from the Bill & Melinda Gates Foundation, the database has been published every three years since 2011. The model is based on the report published in 2017 .there are three model that predicts the financial Technology usage  and factors impacting that ,based on the income economy group a country belongs to .